\name{print.rbugs}
\alias{print.rbugs}
\title{Printing a rbugs object}
\description{Printing a \code{rbugs} object}
\usage{
\method{print}{rbugs}(x, ...)
}
\arguments{
  \item{x}{an object of class 'rbugs', see \code{\link{rbugs}} for details}
  \item{...}{further arguments to \code{\link{print}}}
}
\seealso{\code{\link{rbugs}}}
\keyword{print}
