\name{rbugs2coda}
\alias{rbugs2coda}
\title{ Rbugs to coda }
\description{Convert a \code{rbugs} object to a \code{coda} format}
\usage{
rbugs2coda(model)
}
\arguments{
  \item{model}{an object of class 'rbugs', see \code{\link{rbugs}} for details}
}
\seealso{\code{\link{rbugs}}}
\keyword{coda}
